requirements.txt
langchain
python-dotenv==1.0.0
streamlit==1.22.0
unstructured==0.9.2
tiktoken
faiss-cpu
libmagic==1.0
python-magic==0.4.27
python-magic-bin==0.4.14
OpenAI
nltk
pypdf
requests
math